fix colors when shipping via pusher
When sent via an alternative transport, there are cases where the log lines sent are not properly escaped, and thus we get ugly output on the screen.

To combat this, and to make theming a bit easier, we use css to the elements that are different, and then attach actual styling within the browser.